Update models
Update to include `jax` model. it redefines `evaluate_interp_model` so that its logic agrees with with what is in `modules/spline_redshift.py`.

It also redefines `propose_death_move` such that one cannot get rid of the first knot. That is important because the logic we have for the JAX RJ sampling does not handle extrapolation well. So we need to make sure the first knot is always still in the model.